move tests for ott-common stuff into ott-common
