+++ client/tests/unit/timestamp.spec.ts
@@ -1,7 +1,6 @@
 import { it, describe, expect } from "vitest";
 import dayjs from "dayjs";
 import { secondsToTimestamp, timestampToSeconds } from "../../src/util/timestamp";
-import { calculateCurrentPosition } from "ott-common/timestamp";
 
 describe("secondsToTimestamp spec", () => {
 	it("handles positive values", () => {
@@ -17,53 +16,6 @@ describe("secondsToTimestamp spec", () => {
 	it("handles time spans over an hour", () => {
 		let t = secondsToTimestamp(5400);
 		expect(t).toMatch("01:30:00");
-	});
-});
-
-// TODO: move this to ott-common
-describe("calculateCurrentPosition spec", () => {
-	it("should calculate the correct playback position", () => {
-		expect(calculateCurrentPosition(dayjs(), dayjs(), 0)).toBeCloseTo(0);
-		expect(calculateCurrentPosition(dayjs(), dayjs(), 1)).toBeCloseTo(1);
-		expect(
-			calculateCurrentPosition(
-				dayjs("8 Mar 2020 05:00:00 GMT"),
-				dayjs("8 Mar 2020 05:00:03 GMT"),
-				0
-			)
-		).toBeCloseTo(3);
-		expect(
-			calculateCurrentPosition(
-				dayjs("8 Mar 2020 05:00:00 GMT"),
-				dayjs("8 Mar 2020 05:01:00 GMT"),
-				0
-			)
-		).toBeCloseTo(60);
-		expect(
-			calculateCurrentPosition(
-				dayjs("8 Mar 2020 05:00:00 EST"),
-				dayjs("8 Mar 2020 05:00:03 EST"),
-				0
-			)
-		).toBeCloseTo(3);
-
-		expect(
-			calculateCurrentPosition(
-				dayjs("8 Mar 2020 05:00:00 EST"),
-				dayjs("8 Mar 2020 05:00:03 EST"),
-				0,
-				2
-			)
-		).toBeCloseTo(6);
-
-		expect(
-			calculateCurrentPosition(
-				dayjs("8 Mar 2020 05:00:00 EST"),
-				dayjs("8 Mar 2020 05:00:03 EST"),
-				0,
-				0.5
-			)
-		).toBeCloseTo(1.5);
 	});
 });
